+++ CHANGELOG.md
@@ -5,6 +5,13 @@ file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- Fixed the `request::Handler::trusted_domain_list` function to correctly parse
+  the argument (#39).
+
 ## [0.5.0] - 2024-09-05
 
 ### Added
@@ -207,6 +214,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - `client::handshake` implements the application-level handshake process for the
   client after a QUIC connection is established.
 
+[Unreleased]: https://github.com/petabi/review-protocol/compare/0.5.0...main
 [0.5.0]: https://github.com/petabi/review-protocol/compare/0.4.2...0.5.0
 [0.4.2]: https://github.com/petabi/review-protocol/compare/0.4.1...0.4.2
 [0.4.1]: https://github.com/petabi/review-protocol/compare/0.4.0...0.4.1